--- a/dask/array/routines.py
+++ b/dask/array/routines.py
@@ -17,7 +17,7 @@ from ..highlevelgraph import HighLevelGraph
 from ..utils import funcname, derived_from, is_arraylike
 from . import chunk
 from .creation import arange, diag, empty, indices
-from .utils import safe_wraps, validate_axis, meta_from_array, zeros_like_safe
+from .utils import safe_wraps, validate_axis, zeros_like_safe
 from .wrap import ones
 from .ufunc import multiply, sqrt
 
@@ -642,7 +642,7 @@ def bincount(x, weights=None, minlength=0, split_every=None):
     token = tokenize(x, weights, minlength)
     args = [x, "i"]
     if weights is not None:
-        meta = meta_from_array(weights)
+        meta = np.bincount([1], weights=[1])
         args.extend([weights, "i"])
     else:
         meta = np.bincount([])