- [x] Closes #4852
- [x] Tests added / passed
- [x] Passes `black dask` / `flake8 dask`

This implementation is not faster with this benchmark:

```py
import dask.array as da
import numpy as np

rng = np.random.RandomState(42)
x = rng.randint(5000, size=10_000_000)
x_da = da.from_array(x, chunks=300_000)
out = da.bincount(x_da)
```

```py
%%timeit
_ = out.compute()
```

### This PR

```
26.1 ms ± 2.06 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### Master

```
21.2 ms ± 272 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Maybe there is not a benefit to tree reduction here?