@jsignell 2021-02-17
Huh. If there is no improvement do you think it's worth including this? Or do you think there'd be an improvement with a different benchmark?