@thomasjpfan 2021-02-23
I think the benefit of this PR is memory usage for larger arrays:

```py
import dask.array as da
import dask
from dask.diagnostics import Profiler, ResourceProfiler, CacheProfiler
import numpy as np

rng = np.random.RandomState(42)
x = rng.randint(50000, size=500_000_000)
x_da = da.from_array(x, chunks=100_000)
out = da.bincount(x_da)

with ResourceProfiler(dt=0.01) as rprof:
    out.compute()
    
rprof.visualize()
```

On `master`: Goes to ~ 7.4 GB

![bokeh_plot-2](https://user-images.githubusercontent.com/5402633/108801236-2fda9b80-7563-11eb-90a5-a254493299db.png)

This PR: Stays at ~ 4.7GB

![bokeh_plot](https://user-images.githubusercontent.com/5402633/108801244-3537e600-7563-11eb-9cbe-4a8a1f8ebb48.png)

`master` is always a little faster but it uses more memory.