@jsignell 2021-02-23
Oh that is a huge memory improvement! Thanks for taking the time to write that up :)