@thomasjpfan 2021-02-23
Same benchmark with `Profiler` included:

## On `master`

![Screen Shot 2021-02-23 at 3 56 11 PM](https://user-images.githubusercontent.com/5402633/108907155-bfc62700-75ef-11eb-84e7-fbf55c4bac7f.png)

## This PR

![Screen Shot 2021-02-23 at 3 55 41 PM](https://user-images.githubusercontent.com/5402633/108907170-c5237180-75ef-11eb-8369-6afdb17bda52.png)

My guess is that there is more gil contention when running `_bincount_agg` and `np.bincount` together.